- got correct linear speed pid coeffs - eliminated quantization in pid compute processing (changed ints to floats) - added hysteresis to treeeyes stop function - added wcet images

+++ Iaroslav_Pshenichnikov/tarefa_2/main/tasks/speed_ctrl_task.h
@@ -14,14 +14,14 @@ typedef struct {
 
 
 typedef struct {
-    int L;
-    int R;
+    float L;
+    float R;
 } speed_t;
 // cm/s
 
 typedef struct {
-    int l_wheel;
-    int r_wheel;
+    float l_wheel;
+    float r_wheel;
 } cinematic_t; /// Velocidades lineares das rodas, em cm/s
 
 typedef struct {
